move queuehandler to tq
