+++ dispatch/dispatch.go
@@ -10,7 +10,7 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
+	"github.com/m-lab/etl-gardener/cloud/tq"
 	"google.golang.org/api/option"
 )
 
@@ -21,57 +21,6 @@ import (
 // We don't care which queue a date goes to.  Only reason for multiple queues is to rate
 // limit each individual day.
 
-func init() {
-	// Register the metrics defined with Prometheus's default registry.
-	prometheus.MustRegister(FailCount)
-	prometheus.MustRegister(ErrorCount)
-	prometheus.MustRegister(WarningCount)
-}
-
-var (
-	// ErrorCount counts non-fatal errors.
-	//
-	// Provides metrics:
-	//   etl_error_count{status}
-	// Example usage:
-	// metrics.ErrorCount.WithLabelValues("ok").Inc()
-	ErrorCount = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "etl_error_count",
-			Help: "Number of non-fatal errors.",
-		},
-		[]string{"status"},
-	)
-
-	// WarningCount counts the number of files processed with each status.
-	//
-	// Provides metrics:
-	//   etl_fail_count{status}
-	// Example usage:
-	// metrics.WarningCount.WithLabelValues("funny xyz").Inc()
-	WarningCount = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "etl_warning_count",
-			Help: "Number of processing warnings.",
-		},
-		[]string{"status"},
-	)
-
-	// FailCount counts the number of files processed with each status.
-	//
-	// Provides metrics:
-	//   etl_fail_count{status}
-	// Example usage:
-	// metrics.FailCount.WithLabelValues("BadTableName").Inc()
-	FailCount = prometheus.NewCounterVec(
-		prometheus.CounterOpts{
-			Name: "etl_fail_count",
-			Help: "Number of processing failures.",
-		},
-		[]string{"status"},
-	)
-)
-
 // Dispatcher globals
 type Dispatcher struct {
 	Queues      []chan<- string
@@ -81,7 +30,9 @@ type Dispatcher struct {
 }
 
 // Dispatcher related errors.
-var ErrTerminating = errors.New("dispatcher is terminating")
+var (
+	ErrTerminating = errors.New("dispatcher is terminating")
+)
 
 // NewDispatcher creates a dispatcher that will spread requests across multiple
 // QueueHandlers.
@@ -90,7 +41,7 @@ func NewDispatcher(httpClient *http.Client, project, queueBase string, numQueues
 	queues := make([]chan<- string, 0, numQueues)
 	done := make([]<-chan bool, 0, numQueues)
 	for i := 0; i < numQueues; i++ {
-		q, d, err := NewChannelQueueHandler(httpClient, project,
+		q, d, err := tq.NewChannelQueueHandler(httpClient, project,
 			fmt.Sprintf("%s%d", queueBase, i), bucketOpts...)
 		if err != nil {
 			return nil, err
